+++ src/components/ImageDepViewer.js
@@ -2,7 +2,7 @@ import _ from 'lodash/fp';
 import { Fragment, useEffect, useState } from 'react';
 import { div, h, table, tbody, td, thead, tr } from 'react-hyperscript-helpers';
 import { Select } from 'src/components/common';
-import { fetchOk } from 'src/libs/ajax/ajax-common';
+import { fetchOk } from 'src/libs/ajax/fetch/fetch-core';
 import { withErrorReporting } from 'src/libs/error';
 import { useCancellation } from 'src/libs/react-utils';
