+++ src/libs/ajax/AzureStorage.ts
@@ -2,8 +2,10 @@ import _ from 'lodash/fp';
 import { AnalysisFile, AnalysisFileMetadata } from 'src/analysis/useAnalysisFiles';
 import { AbsolutePath, getDisplayName, getExtension, getFileName } from 'src/analysis/utils/file-utils';
 import { runtimeToolLabels } from 'src/analysis/utils/tool-utils';
+import { authOpts } from 'src/auth/auth-fetch';
 import { Ajax } from 'src/libs/ajax';
-import { authOpts, fetchOk, fetchWorkspaceManager } from 'src/libs/ajax/ajax-common';
+import { fetchWorkspaceManager } from 'src/libs/ajax/ajax-common';
+import { fetchOk } from 'src/libs/ajax/fetch/fetch-core';
 import { getConfig } from 'src/libs/config';
 import * as Utils from 'src/libs/utils';
 import { cloudProviderTypes } from 'src/workspaces/utils';
