+++ src/libs/ajax/Billing.ts
@@ -1,12 +1,8 @@
 import _ from 'lodash/fp';
 import * as qs from 'qs';
-import {
-  authOpts,
-  fetchBillingProfileManager,
-  fetchOrchestration,
-  fetchRawls,
-  jsonBody,
-} from 'src/libs/ajax/ajax-common';
+import { authOpts } from 'src/auth/auth-fetch';
+import { fetchBillingProfileManager, fetchOrchestration, fetchRawls } from 'src/libs/ajax/ajax-common';
+import { jsonBody } from 'src/libs/ajax/fetch/fetch-core';
 import { WorkspacePolicy } from 'src/workspaces/utils';
 
 export interface GoogleBillingAccount {
