+++ src/libs/ajax/Catalog.ts
@@ -1,5 +1,7 @@
 import * as _ from 'lodash/fp';
-import { authOpts, fetchCatalog, jsonBody } from 'src/libs/ajax/ajax-common';
+import { authOpts } from 'src/auth/auth-fetch';
+import { fetchCatalog } from 'src/libs/ajax/ajax-common';
+import { jsonBody } from 'src/libs/ajax/fetch/fetch-core';
 
 // Types are pulled from https://github.com/DataBiosphere/terra-data-catalog/blob/main/common/src/main/resources/schema/development/schema.json
 export type StorageSystem = 'wks' | 'ext' | 'tdr';
