+++ src/libs/ajax/GoogleStorage.ts
@@ -9,16 +9,10 @@ import {
   ToolLabel,
 } from 'src/analysis/utils/tool-utils';
 import { getAuthToken } from 'src/auth/auth';
-import {
-  authOpts,
-  checkRequesterPaysError,
-  fetchOk,
-  fetchSam,
-  jsonBody,
-  withRetryOnError,
-  withUrlPrefix,
-} from 'src/libs/ajax/ajax-common';
+import { authOpts } from 'src/auth/auth-fetch';
+import { checkRequesterPaysError, fetchSam, withRetryOnError } from 'src/libs/ajax/ajax-common';
 import { canUseWorkspaceProject } from 'src/libs/ajax/Billing';
+import { fetchOk, jsonBody, withUrlPrefix } from 'src/libs/ajax/fetch/fetch-core';
 import { getConfig } from 'src/libs/config';
 import { knownBucketRequesterPaysStatuses, requesterPaysProjectStore, workspaceStore } from 'src/libs/state';
 import * as Utils from 'src/libs/utils';
