+++ src/libs/ajax/ajax-common.ts
@@ -1,28 +1,12 @@
 import { delay } from '@terra-ui-packages/core-utils';
 import _ from 'lodash/fp';
-import {
-  AuthTokenState,
-  getAuthToken,
-  getAuthTokenFromLocalStorage,
-  loadAuthToken,
-  sendRetryMetric,
-} from 'src/auth/auth';
-import { sessionExpirationErrorMessage } from 'src/auth/auth-errors';
-import { signOut, SignOutCause } from 'src/auth/signout/sign-out';
-import { FetchFn } from 'src/libs/ajax/data-client-common';
-import { withCancellation, withErrorRejection, withInstrumentation } from 'src/libs/ajax/fetch/fetch-core';
+import { withRetryAfterReloadingExpiredAuthToken } from 'src/auth/auth-fetch';
 import { getConfig } from 'src/libs/config';
 
-export const authOpts = (token = getAuthToken()) => ({ headers: { Authorization: `Bearer ${token}` } });
-export const jsonBody = (body) => ({
-  body: JSON.stringify(body),
-  headers: { 'Content-Type': 'application/json' },
-});
-export const appIdentifier = { headers: { 'X-App-ID': 'Saturn' } };
+import { FetchFn } from './data-client-common';
+import { fetchOk, withCancellation, withInstrumentation, withUrlPrefix } from './fetch/fetch-core';
 
-export const withUrlPrefix = _.curry((prefix, wrappedFetch) => (path, ...args) => {
-  return wrappedFetch(prefix + path, ...args);
-});
+export const appIdentifier = { headers: { 'X-App-ID': 'Saturn' } };
 
 export const withRetryOnError = _.curry((shouldNotRetryFn, wrappedFetch) => async (...args) => {
   const timeout = 5000;
@@ -71,70 +55,10 @@ export async function makeRequestRetry(request: Function, retryCount: number, ti
   }
 }
 
-const isUnauthorizedResponse = (error: unknown): boolean => error instanceof Response && error.status === 401;
-
-const createRequestWithStoredAuthToken = async (
-  wrappedFetch: FetchFn,
-  resource: RequestInfo | URL,
-  options?: RequestInit
-): Promise<Response> => {
-  const localToken = (await getAuthTokenFromLocalStorage())!;
-  return createRequestWithNewAuthToken(localToken, wrappedFetch, resource, options);
-};
-
-const createRequestWithNewAuthToken = (
-  token: string,
-  wrappedFetch: FetchFn,
-  resource: RequestInfo | URL,
-  options?: RequestInit
-): Promise<Response> => {
-  const optionsWithNewAuthToken = _.merge(options, authOpts(token));
-  return wrappedFetch(resource, optionsWithNewAuthToken);
-};
-
-export const withRetryAfterReloadingExpiredAuthToken =
-  (wrappedFetch: FetchFn): FetchFn =>
-  async (resource: RequestInfo | URL, options?: RequestInit): Promise<Response> => {
-    const preRequestAuthToken = getAuthToken();
-    const requestHasAuthHeader = _.isMatch(authOpts(), options as object);
-    try {
-      if (requestHasAuthHeader) {
-        // use auth token in local storage
-        return await createRequestWithStoredAuthToken(wrappedFetch, resource, options);
-      }
-      return await wrappedFetch(resource, options);
-    } catch (error: unknown) {
-      if (isUnauthorizedResponse(error) && requestHasAuthHeader) {
-        const reloadedAuthTokenState: AuthTokenState = await loadAuthToken();
-        const postRequestAuthToken = getAuthToken();
-        if (reloadedAuthTokenState.status === 'success') {
-          sendRetryMetric();
-          return await createRequestWithStoredAuthToken(wrappedFetch, resource, options);
-        }
-        // if the auth token the request was made with does not match the current auth token
-        // that means that the user has already been signed out and signed in again to receive a new token
-        // in this case, we should not sign the user out again
-        if (preRequestAuthToken === postRequestAuthToken) {
-          const signOutCause: SignOutCause = 'errorRefreshingAuthToken';
-          signOut(signOutCause);
-        }
-        throw new Error(sessionExpirationErrorMessage);
-      } else {
-        throw error;
-      }
-    }
-  };
-
 export const withAppIdentifier =
   (wrappedFetch: FetchFn): FetchFn =>
   (url, options) => {
     return wrappedFetch(url, _.merge(options, appIdentifier));
-  };
-
-export const withAuthSession =
-  (wrappedFetch: FetchFn): FetchFn =>
-  (url, options) => {
-    return wrappedFetch(url, _.merge(options, authOpts()));
   };
 
 export type RequesterPaysErrorInfo = {
@@ -160,8 +84,6 @@ export const checkRequesterPaysError = async (response): Promise<RequesterPaysEr
 export const responseContainsRequesterPaysError = (responseText) => {
   return _.includes('requester pays', responseText);
 };
-
-export const fetchOk = _.flow(withInstrumentation, withCancellation, withErrorRejection)(fetch);
 
 export const fetchLeo = _.flow(
   withUrlPrefix(`${getConfig().leoUrlRoot}/`),
@@ -220,10 +142,6 @@ export const fetchDrsHub = _.flow(
   withUrlPrefix(`${getConfig().drsHubUrlRoot}/`),
   withRetryAfterReloadingExpiredAuthToken
 )(fetchOk);
-
-// Don't wrap Bard calls in withRetryAfterReloadingExpiredAuthToken, because
-// that wrapper itself generates metrics.
-export const fetchBard = withUrlPrefix(`${getConfig().bardRoot}/`, fetchOk);
 
 export const fetchEcm = _.flow(
   withUrlPrefix(`${getConfig().externalCreds?.urlRoot}/`),
