+++ src/libs/ajax/leonardo/Disks.ts
@@ -1,8 +1,9 @@
+import { jsonBody } from '@terra-ui-packages/data-client-core';
 import { AuditInfo, CloudContext, LeoResourceLabels } from '@terra-ui-packages/leonardo-data-client';
 import _ from 'lodash/fp';
 import * as qs from 'qs';
-import { appIdentifier, authOpts, fetchLeo, withAppIdentifier, withAuthSession } from 'src/libs/ajax/ajax-common';
-
+import { authOpts } from 'src/auth/auth-fetch';
+import { appIdentifier, fetchLeo, withAppIdentifier, withAuthSession } from 'src/libs/ajax/ajax-common';
 import { FetchFn } from '../data-client-common';
 import { LeoDisksV1DataClient, makeLeoDisksV1DataClient } from './LeoDisksV1DataClient';
 import { LeoDisksV2DataClient, makeLeoDisksV2DataClient } from './LeoDisksV2DataClient';
