+++ src/libs/ajax/leonardo/Runtimes.ts
@@ -1,18 +1,16 @@
 import _ from 'lodash/fp';
 import * as qs from 'qs';
 import { version } from 'src/analysis/utils/gce-machines';
+import { authOpts, withAuthSession } from 'src/auth/auth-fetch';
 import {
   appIdentifier,
-  authOpts,
   DEFAULT_RETRY_COUNT,
   DEFAULT_TIMEOUT_DURATION,
   fetchLeo,
-  fetchOk,
-  jsonBody,
   makeRequestRetry,
   withAppIdentifier,
-  withAuthSession,
 } from 'src/libs/ajax/ajax-common';
+import { fetchOk, jsonBody } from 'src/libs/ajax/fetch/fetch-core';
 import { LeoRuntimesV1DataClient, makeLeoRuntimesV1DataClient } from 'src/libs/ajax/leonardo/LeoRuntimesV1DataClient';
 import { RawRuntimeConfig } from 'src/libs/ajax/leonardo/models/api-runtime-config';
 import {
