+++ src/libs/ajax/workflows-app/CbasPact.test.js
@@ -2,7 +2,8 @@ import 'setimmediate';
 
 import { MatchersV3, PactV3, SpecificationVersion } from '@pact-foundation/pact';
 import path from 'path';
-import { fetchFromProxy, fetchOk, jsonBody } from 'src/libs/ajax/ajax-common';
+import { fetchFromProxy } from 'src/libs/ajax/ajax-common';
+import { fetchOk, jsonBody } from 'src/libs/ajax/fetch/fetch-core';
 import { Cbas } from 'src/libs/ajax/workflows-app/Cbas';
 import {
   runSetInputDef,
@@ -14,7 +15,15 @@ import {
 jest.mock('src/libs/ajax/ajax-common', () => ({
   ...jest.requireActual('src/libs/ajax/ajax-common'),
   fetchFromProxy: jest.fn(),
+}));
+
+jest.mock('src/libs/ajax/fetch/fetch-core', () => ({
+  ...jest.requireActual('src/libs/ajax/fetch/fetch-core'),
   fetchOk: jest.fn(),
+}));
+
+jest.mock('src/auth/auth-fetch', () => ({
+  ...jest.requireActual('src/auth/auth-fetch'),
   authOpts: jest.fn(),
 }));
